Move custom IME shadow into widget package
